Delta version (play vs computer)
Finally added a second mode of the game: let the computer guess the
sequence of your creation. As of now it is blindly hitting the colors
until it hits Bs, where the color would sort of 'stick' in a guessed
sequence. It allows a player to win sometimes, given the 12 turns rule.